C# profiler overlay Added generic key/value pairs to EditorSettings Fixed an issue with VirtualInput where toggle states would reset immediately

+++ SBansheeEngine/Include/BsScriptTime.h
@@ -13,7 +13,7 @@ namespace BansheeEngine
 	private:
 		static float internal_getElapsed();
 		static float internal_getFrameDelta();
-		static UINT32 internal_getFrameNumber();
+		static UINT64 internal_getFrameNumber();
 		static UINT64 internal_getPrecise();
 
 		ScriptTime(MonoObject* instance);
